+++ python/aistore/sdk/etl/webserver/__init__.py
@@ -3,5 +3,7 @@
 #
 
 from aistore.sdk.etl.webserver.http_multi_threaded_server import HTTPMultiThreadedServer
+from aistore.sdk.etl.webserver.fastapi_server import FastAPIServer
 
-__all__ = ["HTTPMultiThreadedServer"]
+
+__all__ = ["HTTPMultiThreadedServer", "FastAPIServer"]